Rename store from "gameState" to "gameStore"
