+++ game/store/hooks.ts
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector } from "react-redux";
-import type { GameDispatch, RootGameState } from "./gameState";
+import type { GameDispatch, RootGameState } from "./gameStore";
 
 // Typed versions of plain `useDispatch` and `useSelector`
 export const useGameDispatch = useDispatch.withTypes<GameDispatch>();
