+++ game/store/index.ts
@@ -1,5 +1,5 @@
-import gameState from "./gameState";
+import gameStore from "./gameStore";
 
 export { deleteLetterFromGuess, selectGuess, updateGuess } from "./guessSlice";
 export { useGameDispatch, useGameSelector } from "./hooks";
-export default gameState;
+export default gameStore;
